Refactor shared logic to trait.
